django_wikinya/requirements.txt
django >= 1.3
South
PIL
PyYAML

django-extensions
django-compress

django-reversion

django-mptt

python-creole
werkzeug

